feat: [RTD-785] add file report feature (#307)
* [RTD-839] add data model for the file report (#284)

* [RTD-835] define feign client method (#285)

* [RTD-836] implement rest client (#286)

* [RTD-1045] define file report directory (#304)

* [RTD-837] add step to save reports (#305)

* [RTD-838] add file report acceptance test (#306)

* [RTD-1105] add docs on methods (#308)

* [RTD-1106] update pom version to 1.5.0 (#309)

+++ api/batch/src/main/java/it/gov/pagopa/rtd/transaction_filter/batch/step/processor/TransactionAggregationWriterProcessor.java
@@ -35,7 +35,7 @@ public class TransactionAggregationWriterProcessor implements ItemProcessor<Aggr
         aggregate.setTransmissionDate(transmissionDate);
         aggregate.setAccountingDate(key.getAccountingDate().getDate());
         aggregate.setNumTrx(storeService.getAggregate(key).getNumTrx());
-        aggregate.setTotalAmount((long) storeService.getAggregate(key).getTotalAmount());
+        aggregate.setTotalAmount(storeService.getAggregate(key).getTotalAmount());
         aggregate.setCurrency("978");
         aggregate.setAcquirerId(key.getAcquirerId().getId());
         aggregate.setMerchantId(key.getMerchantId());
